refactor: extract ConfirmButton
