+++ apps/cowswap-frontend/src/modules/trade/pure/TradeConfirmation/index.tsx
@@ -1,10 +1,7 @@
-import { ReactElement, useEffect, useRef, useState } from 'react'
+import { ReactElement, ReactNode, useEffect, useRef, useState } from 'react'
 
-import { BackButton, BannerOrientation, ButtonPrimary, ButtonSize, CenteredDots, LongLoadText } from '@cowprotocol/ui'
+import { BackButton, BannerOrientation } from '@cowprotocol/ui'
 
-import { Trans } from '@lingui/macro'
-
-import { upToMedium, useMediaQuery } from 'legacy/hooks/useMediaQuery'
 import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 
 import type { AppDataInfo } from 'modules/appData'
@@ -13,6 +10,7 @@ import { OrderHooksDetails } from 'common/containers/OrderHooksDetails'
 import { CurrencyAmountPreview, CurrencyPreviewInfo } from 'common/pure/CurrencyInputPanel'
 import { CustomRecipientWarningBanner } from 'common/pure/CustomRecipientWarningBanner'
 
+import { ConfirmButton } from './ConfirmButton'
 import { QuoteCountdown } from './CountDown'
 import { useIsPriceChanged } from './hooks/useIsPriceChanged'
 import * as styledEl from './styled'
@@ -36,15 +34,14 @@ export interface TradeConfirmationProps {
   title: ReactElement | string
   isPriceStatic?: boolean
   recipient?: string | null
-  buttonText?: React.ReactNode
+  buttonText?: ReactNode
   children?: (restContent: ReactElement) => ReactElement
 }
 
 // TODO: Break down this large function into smaller functions
-// TODO: Add proper return type annotation
 // TODO: Reduce function complexity by extracting logic
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type, complexity
-export function TradeConfirmation(props: TradeConfirmationProps) {
+// eslint-disable-next-line max-lines-per-function, complexity
+export function TradeConfirmation(props: TradeConfirmationProps): ReactNode {
   const { pendingTrade, forcePriceConfirmation } = useTradeConfirmState()
 
   const propsRef = useRef(props)
@@ -70,17 +67,11 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
     appData,
   } = frozenProps || props
 
-  const [isConfirmClicked, setIsConfirmClicked] = useState(false)
-
   /**
    * Once user sends a transaction, we keep the confirmation content frozen
    */
   useEffect(() => {
     setFrozenProps(hasPendingTrade ? propsRef.current : null)
-
-    if (!hasPendingTrade) {
-      setIsConfirmClicked(false)
-    }
   }, [hasPendingTrade])
 
   const showRecipientWarning =
@@ -93,30 +84,11 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
 
   const { isPriceChanged, resetPriceChanged } = useIsPriceChanged(inputAmount, outputAmount, forcePriceConfirmation)
 
-  const isButtonDisabled =
-    isConfirmDisabled || (isPriceChanged && !isPriceStatic) || hasPendingTrade || isConfirmClicked
+  const isButtonDisabled = isConfirmDisabled || (isPriceChanged && !isPriceStatic) || hasPendingTrade
 
   useEffect(() => {
     window.scrollTo(0, 0)
   }, [])
-
-  const isUpToMedium = useMediaQuery(upToMedium)
-
-  // Combine local onClick logic with incoming onClick
-  // TODO: Add proper return type annotation
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  const handleConfirmClick = async () => {
-    if (isUpToMedium) {
-      window.scrollTo({ top: 0, left: 0 })
-    }
-
-    setIsConfirmClicked(true)
-    const isConfirmed = await onConfirm()
-
-    if (!isConfirmed) {
-      setIsConfirmClicked(false)
-    }
-  }
 
   const hookDetailsElement = (
     <>
@@ -159,15 +131,12 @@ export function TradeConfirmation(props: TradeConfirmationProps) {
 
         {showRecipientWarning && <CustomRecipientWarningBanner orientation={BannerOrientation.Horizontal} />}
         {isPriceChanged && !isPriceStatic && <PriceUpdatedBanner onClick={resetPriceChanged} />}
-        <ButtonPrimary onClick={handleConfirmClick} disabled={isButtonDisabled} buttonSize={ButtonSize.BIG}>
-          {hasPendingTrade || isConfirmClicked ? (
-            <LongLoadText fontSize={15} fontWeight={500}>
-              Confirm with your wallet <CenteredDots smaller />
-            </LongLoadText>
-          ) : (
-            <Trans>{buttonText}</Trans>
-          )}
-        </ButtonPrimary>
+        <ConfirmButton
+          onConfirm={onConfirm}
+          buttonText={buttonText}
+          isButtonDisabled={isButtonDisabled}
+          hasPendingTrade={hasPendingTrade}
+        />
       </styledEl.ContentWrapper>
     </styledEl.WidgetWrapper>
   )
